XIVY-15943 move `useLanguage` to its own file
