+++ packages/cms-editor/src/CmsEditor.tsx
@@ -1,9 +1,8 @@
-import type { CmsEditorDataContext, ContentObject, EditorProps } from '@axonivy/cms-editor-protocol';
+import type { ContentObject, EditorProps } from '@axonivy/cms-editor-protocol';
 import { Flex, PanelMessage, ResizableHandle, ResizablePanel, ResizablePanelGroup, Spinner, useHotkeys } from '@axonivy/ui-components';
 import { IvyIcons } from '@axonivy/ui-icons';
 import { useQuery } from '@tanstack/react-query';
-import i18next from 'i18next';
-import { useEffect, useMemo, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import './CmsEditor.css';
 import { AppProvider } from './context/AppContext';
@@ -13,8 +12,8 @@ import { MainContent } from './main/MainContent';
 import { MainToolbar } from './main/MainToolbar';
 import { useClient } from './protocol/ClientContextProvider';
 import { useAction } from './protocol/use-action';
-import { useMeta } from './protocol/use-meta';
 import { useQueryKeys } from './query/query-client';
+import { useLanguage } from './use-language';
 import { useKnownHotkeys } from './utils/hotkeys';
 
 function CmsEditor(props: EditorProps) {
@@ -104,24 +103,4 @@ export const toolbarTitles = (pmv: string, contentObject?: ContentObject) => {
     detailTitle += ` - ${contentObject.uri.substring(lastSlashIndex + 1)}`;
   }
   return { mainTitle, detailTitle };
-};
-
-export const useLanguage = (context: CmsEditorDataContext) => {
-  const clientLanguageTag = i18next.language;
-  const languageDisplayName = useMemo(() => new Intl.DisplayNames([clientLanguageTag], { type: 'language' }), [clientLanguageTag]);
-
-  const locales = useMeta('meta/locales', context, []);
-  const defaultLanguageTag = useMemo(() => defaultLanguage(locales.data, clientLanguageTag), [locales.data, clientLanguageTag]);
-
-  return { defaultLanguageTag, languageDisplayName };
-};
-
-const defaultLanguage = (locales: Array<string>, clientLanguageTag: string) => {
-  if (locales.includes(clientLanguageTag) || locales.length === 0) {
-    return clientLanguageTag;
-  }
-  if (locales.includes('en')) {
-    return 'en';
-  }
-  return locales[0];
 };
